+++ src/main/java/com/mmo/server/core/math/Vertex.java
@@ -11,18 +11,18 @@ import lombok.ToString;
 @ToString
 public class Vertex {
 
-    protected float x;
-    protected float z;
+    protected int x;
+    protected int z;
 
-    public float getDistance(Vertex other) {
+    public int getDistance(Vertex other) {
         return (x - other.getX()) * (x - other.getX()) + (z - other.getZ()) * (z - other.getZ());
     }
 
     public boolean isNearby(Vertex position, int ratio) {
-        float minX = x - ratio;
-        float maxX = x + ratio;
-        float minZ = z - ratio;
-        float maxZ = z + ratio;
+        int minX = x - ratio;
+        int maxX = x + ratio;
+        int minZ = z - ratio;
+        int maxZ = z + ratio;
 
         return position.getX() >= minX && position.getX() <= maxX
                 && position.getZ() >= minZ && position.getZ() <= maxZ;
